ChatModels/hugging_faceapi.py
from langchain_huggingface import ChatHuggingFace, HuggingFaceEndpoint
from dotenv import load_dotenv

load_dotenv()

llm = HuggingFaceEndpoint(repo_id="meta-llama/Llama-3.1-8B-Instruct", task="text-generation")

model = ChatHuggingFace(llm=llm)  # Example model name

result = model.invoke("What is Narendra Modi?")

print(result.content)

